+++ vitest.config.ts
@@ -5,5 +5,10 @@ export default defineConfig({
     globals: true,
     environment: 'node',
     setupFiles: ['./tests/setup.ts'],
+    coverage: {
+      all: true,
+      include: ['src/**/*.ts'],
+      exclude: ['src/index.ts'] // Usually exclude bootstrap index
+    }
   },
 });